requirements.txt
python-dotenv==1.0.0
anthropic==0.3.11
openai==0.28.1
colorama==0.4.6
requests==2.31.0
boto3==1.28.63
nltk==3.8.1